refactor: 정렬 오래된순 최신순 버그 수정

+++ data/library/src/main/java/com/into/websoso/data/library/mediator/FilteredNovelRemoteMediator.kt
@@ -2,8 +2,14 @@ package com.into.websoso.data.library.mediator
 
 import androidx.paging.ExperimentalPagingApi
 import androidx.paging.LoadType
+import androidx.paging.LoadType.APPEND
+import androidx.paging.LoadType.PREPEND
+import androidx.paging.LoadType.REFRESH
 import androidx.paging.PagingState
 import androidx.paging.RemoteMediator
+import androidx.paging.RemoteMediator.MediatorResult.Error
+import androidx.paging.RemoteMediator.MediatorResult.Success
+import com.into.websoso.core.common.extensions.isCloseTo
 import com.into.websoso.core.database.entity.InDatabaseFilteredNovelEntity
 import com.into.websoso.data.library.datasource.FilteredLibraryLocalDataSource
 import com.into.websoso.data.library.datasource.LibraryRemoteDataSource
@@ -17,23 +23,17 @@ class FilteredNovelRemoteMediator(
     private val sortCriteria: String,
     private val readStatuses: List<String>,
     private val attractivePoints: List<String>,
-    private val novelRating: Float?,
+    private val novelRating: Float,
 ) : RemoteMediator<Int, InDatabaseFilteredNovelEntity>() {
     override suspend fun load(
         loadType: LoadType,
         state: PagingState<Int, InDatabaseFilteredNovelEntity>,
     ): MediatorResult {
         val lastUserNovelId = when (loadType) {
-            LoadType.REFRESH -> null
-            LoadType.PREPEND -> return MediatorResult.Success(endOfPaginationReached = true)
-            LoadType.APPEND -> {
-                when (sortCriteria) {
-                    "RECENT" -> state.lastItemOrNull()?.userNovelId
-                    "OLD" -> state.firstItemOrNull()?.userNovelId
-                    else -> null
-                } ?: return MediatorResult.Success(true)
-            }
-        } ?: DEFAULT_LAST_USER_NOVEL_ID
+            REFRESH -> DEFAULT_LAST_USER_NOVEL_ID
+            PREPEND -> return Success(endOfPaginationReached = true)
+            APPEND -> state.lastItemOrNull()?.userNovelId ?: return Success(true)
+        }
 
         return try {
             val response = libraryRemoteDataSource.getUserNovels(
@@ -44,23 +44,23 @@ class FilteredNovelRemoteMediator(
                 isInterest = if (!isInterested) null else true,
                 readStatuses = readStatuses.ifEmpty { null },
                 attractivePoints = attractivePoints.ifEmpty { null },
-                novelRating = if (novelRating == 0f) null else novelRating,
+                novelRating = if (novelRating.isCloseTo(DEFAULT_NOVEL_RATING)) null else novelRating,
                 query = null,
                 updatedSince = null,
             )
 
-            if (loadType == LoadType.REFRESH) {
-                filteredLibraryLocalDataSource.deleteAllNovels()
-            }
+            if (loadType == REFRESH) filteredLibraryLocalDataSource.deleteAllNovels()
+
             filteredLibraryLocalDataSource.insertNovels(response.userNovels)
 
-            MediatorResult.Success(endOfPaginationReached = !response.isLoadable)
+            Success(endOfPaginationReached = !response.isLoadable)
         } catch (e: Exception) {
-            MediatorResult.Error(e)
+            Error(e)
         }
     }
 
     companion object {
         private const val DEFAULT_LAST_USER_NOVEL_ID = 0L
+        private const val DEFAULT_NOVEL_RATING = 0f
     }
 }
